+++ lib/generators/rpush_generator.rb
@@ -20,8 +20,10 @@ class RpushGenerator < Rails::Generators::Base
       add_rpush_migration('add_adm')
       add_rpush_migration('rename_rapns_to_rpush')
       add_rpush_migration('add_fail_after_to_rpush_notifications')
+      add_rpush_migration('add_processing_to_rpush_notifications')
     else
       add_rpush_migration('add_rpush')
+      add_rpush_migration('add_processing_to_rpush_notifications')
     end
   end