+++ lib/generators/templates/add_processing_to_rpush_notifications.rb
@@ -1,0 +1,9 @@
+class AddProcessingToRpushNotifications < ActiveRecord::Migration
+  def self.up
+    add_column :rpush_notifications, :processing, :boolean, null: false, default: false
+  end
+
+  def self.down
+    remove_column :rpush_notifications, :processing
+  end
+end